fix(fetch): split Request constructors into request_ctor module (file-size lint)
`fetch/mod.rs` was at the 2,000-line lint cap; the #5458 runtime helper pushed
it to 2089. Move both Request constructors (`js_request_new` and the new
`js_request_new_from_init`) into a `request_ctor` sibling module, re-exported
from mod.rs, mirroring the existing `headers`/`dispatch`/`body_metadata` split.
mod.rs is back to 1940 lines. No behavior change; added the missing `# Safety`
doc sections on the two moved FFI fns.
